mengganti double menjadi float

+++ insertion sort/insertion sort.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 using namespace std;
 
-double arr[20];            //Membuat array dengan panjang data 20
-double n;                 //Membuat Variable inputan n
+float arr[20];            //Membuat array dengan panjang data 20
+float n;                 //Membuat Variable inputan n
 
 void input() {          //Procedure Input
     while (true) {
